app/src/main/java/com/example/my_traveler_app/ui/login/LoginResult.java
package com.example.my_traveler_app.ui.login;

import androidx.annotation.Nullable;

/**
 * Authentication result : success (user details) or error message.
 */
public class LoginResult {
    @Nullable
    private LoggedInUserView success;
    @Nullable
    private Object error;

    LoginResult(@Nullable Object error) {
        this.error = error;
    }

    LoginResult(@Nullable LoggedInUserView success) {
        this.success = success;
    }

    @Nullable
    public LoggedInUserView getSuccess() {
        return success;
    }

    @Nullable
    public Object getError() {
        return error;
    }
}